refactor(admin): remove dead and pass-through code
