+++ apps/admin/src/app/(protected)/(finance)/finance/page.tsx
@@ -3,7 +3,6 @@ import { mapErrorToMessage } from '@/lib/mapErrorToMessage';
 
 import { useCallback, useEffect, useState } from "react";
 import { usePathname, useRouter, useSearchParams } from "next/navigation";
-import { ColumnDef } from "@/components/ui/DataTable";
 import { Transaction, FinanceStats } from "@/types/transaction";
 import { fetchFinanceStats, fetchFinanceTransactions } from "@/lib/api/finance";
 import {
@@ -15,13 +14,14 @@ import {
     TrendingUp,
     Calendar,
     Wallet,
+    DataTable,
+    AlertCircle,
+    type ColumnDef,
 } from "@esparex/ui";
 import { DashboardCard } from "@/components/dashboard/DashboardCard";
 import { AdminPageShell } from "@/components/layout/AdminPageShell";
 import { AdminModuleTabs } from "@/components/layout/AdminModuleTabs";
 import { financeTabs } from "@/components/layout/adminModuleTabSets";
-import { DataTable } from "@/components/ui/DataTable";
-import { AlertCircle } from "@esparex/ui";
 import {
     buildUrlWithSearchParams,
     normalizeSearchParamValue,
