+++ apps/admin/src/app/(protected)/(finance)/plans/page.tsx
@@ -2,8 +2,7 @@
 
 import { useEffect, useState } from "react";
 import { usePathname, useRouter, useSearchParams } from "next/navigation";
-import { ColumnDef } from "@/components/ui/DataTable";
-import { Plan } from "@esparex/contracts";
+import type { Plan } from "@esparex/contracts";
 import {
     CreditCard,
     Filter,
@@ -16,6 +15,9 @@ import {
     Archive,
     RotateCcw,
     ShieldCheck,
+    DataTable,
+    AlertCircle,
+    type ColumnDef,
 } from "@esparex/ui";
 import { PlanFormModal } from "@/components/plans/PlanFormModal";
 import { ArchivePlanModal } from "@/components/plans/ArchivePlanModal";
@@ -23,8 +25,6 @@ import { AdminPageShell } from "@/components/layout/AdminPageShell";
 import { AdminModuleTabs } from "@/components/layout/AdminModuleTabs";
 import { AdminFilterToolbar } from "@/components/layout/AdminFilterToolbar";
 import { financeTabs } from "@/components/layout/adminModuleTabSets";
-import { DataTable } from "@/components/ui/DataTable";
-import { AlertCircle } from "@esparex/ui";
 import { ConfirmDeactivateDialog } from "@/components/finance/ConfirmDeactivateDialog";
 import {
     buildUrlWithSearchParams,
