+++ apps/admin/src/app/(protected)/(system)/admin-sessions/page.tsx
@@ -1,16 +1,14 @@
 "use client";
 
 import { useEffect, useState } from "react";
-import { DataTable, type ColumnDef } from "@/components/ui/DataTable";
-import type { AdminSessionItem } from "@/types/adminSession";
-import { Power, AlertTriangle, Loader2, AlertCircle } from "@esparex/ui";
+import { DataTable, type ColumnDef, StatusChip, Power, AlertTriangle, Loader2, AlertCircle } from "@esparex/ui";
 import { AdminPageShell } from "@/components/layout/AdminPageShell";
 import { AdminModuleTabs } from "@/components/layout/AdminModuleTabs";
 import { administrationTabs } from "@/components/layout/adminModuleTabSets";
-import { StatusChip } from "@/components/ui/StatusChip";
 import { AdminFilterToolbar } from "@/components/layout/AdminFilterToolbar";
 import { CatalogModal } from "@/components/catalog/CatalogModal";
 import { useAdminSessions } from "@/hooks/useAdminSessions";
+import type { AdminSessionItem } from "@/types/adminSession";
 
 export default function AdminSessionsPage() {
     const {
