--- apps/admin/src/app/(protected)/businesses/BusinessesView.tsx
@@ -1,1 +1,0 @@
-export { default } from './businesses-view/main';
