+++ apps/admin/src/app/(protected)/businesses/page.tsx
@@ -1,4 +1,4 @@
-import BusinessesView from "./BusinessesView";
+import BusinessesView from "./businesses-view/main";
 
 export const metadata = {
   title: "Businesses | Esparex Admin",
