+++ apps/admin/src/app/(protected)/reports/page.tsx
@@ -2,12 +2,11 @@
 
 import { useEffect, useMemo, useState } from "react";
 import { usePathname, useRouter, useSearchParams } from "next/navigation";
-import { AlertCircle, CheckCircle2, Eye, ShieldAlert, XCircle } from "@esparex/ui";
+import { AlertCircle, CheckCircle2, Eye, ShieldAlert, XCircle, DataTable, type ColumnDef } from "@esparex/ui";
 import { AdminPageShell } from "@/components/layout/AdminPageShell";
 import { AdminModuleTabs } from "@/components/layout/AdminModuleTabs";
 import { AdminFilterToolbar } from "@/components/layout/AdminFilterToolbar";
 import { AdminActionMenu } from "@/components/layout/AdminActionMenu";
-import { DataTable, type ColumnDef } from "@/components/ui/DataTable";
 import { ADMIN_UI_ROUTES, readPositiveIntParam, readStringParam } from "@/lib/adminUiRoutes";
 import { useModerationReports, type ReportQueueItem } from "@/hooks/useModerationReports";
