--- apps/admin/src/components/catalog/tabs/ModelsTab.tsx
@@ -1,1 +1,0 @@
-export { default } from "./models-tab/main";
