--- apps/admin/src/components/ui/DataTable.tsx
@@ -1,1 +1,0 @@
-export { DataTable, type ColumnDef } from "@esparex/ui";
