--- apps/admin/src/components/ui/StatusChip.tsx
@@ -1,1 +1,0 @@
-export { StatusChip } from "@esparex/ui";
