--- apps/admin/src/lib/utils/idUtils.ts
@@ -1,4 +1,0 @@
-import { sanitizeMongoObjectId, normalizeObjectIdLike } from "@esparex/shared";
-
-export { sanitizeMongoObjectId, normalizeObjectIdLike };
-
