--- apps/admin/src/types/screenSize.ts
@@ -1,1 +1,0 @@
-export type { ScreenSize } from "@esparex/contracts";
